+++ packages/desktop-client/src/components/mobile/budget/BudgetTable.jsx
@@ -1,12 +1,5 @@
 import React, { memo, useCallback, useRef } from 'react';
-import {
-  DropIndicator,
-  ListBox,
-  ListBoxItem,
-  useDragAndDrop,
-} from 'react-aria-components';
 import { useTranslation } from 'react-i18next';
-import { useListData } from 'react-stately';
 
 import { Button } from '@actual-app/components/button';
 import { Card } from '@actual-app/components/card';
@@ -24,7 +17,6 @@ import {
   trackingBudget,
   uncategorizedCount,
 } from 'loot-core/client/queries';
-import { moveCategory } from 'loot-core/client/queries/queriesSlice';
 import * as monthUtils from 'loot-core/shared/months';
 import { groupById, integerToCurrency } from 'loot-core/shared/util';
 
@@ -33,7 +25,6 @@ import { useFeatureFlag } from '../../../hooks/useFeatureFlag';
 import { useLocale } from '../../../hooks/useLocale';
 import { useLocalPref } from '../../../hooks/useLocalPref';
 import { useNavigate } from '../../../hooks/useNavigate';
-import { useNotes } from '../../../hooks/useNotes';
 import { useSyncedPref } from '../../../hooks/useSyncedPref';
 import { useUndo } from '../../../hooks/useUndo';
 import { SvgLogo } from '../../../icons/logo';
@@ -59,15 +50,21 @@ import { useSheetValue } from '../../spreadsheet/useSheetValue';
 import { MOBILE_NAV_HEIGHT } from '../MobileNavTabs';
 import { PullToRefresh } from '../PullToRefresh';
 
+import { BudgetCell } from './BudgetCell';
+import { IncomeCategoryList } from './IncomeCategoryList';
 import { ListItem } from './ListItem';
 
-const PILL_STYLE = {
+export const PILL_STYLE = {
   borderRadius: 16,
   color: theme.pillText,
   backgroundColor: theme.pillBackgroundLight,
 };
 
-function getColumnWidth({ show3Cols, isSidebar = false, offset = 0 } = {}) {
+export function getColumnWidth({
+  show3Cols,
+  isSidebar = false,
+  offset = 0,
+} = {}) {
   // If show3Cols = 35vw | 20vw | 20vw | 20vw,
   // Else = 45vw | 25vw | 25vw,
   if (!isSidebar) {
@@ -225,136 +222,6 @@ function Saved({ projected, onPress, show3Cols }) {
         />
       </Button>
     </View>
-  );
-}
-
-function BudgetCell({
-  name,
-  binding,
-  style,
-  category,
-  month,
-  onBudgetAction,
-  children,
-  ...props
-}) {
-  const { t } = useTranslation();
-  const columnWidth = getColumnWidth();
-  const dispatch = useDispatch();
-  const format = useFormat();
-  const { showUndoNotification } = useUndo();
-  const [budgetType = 'rollover'] = useSyncedPref('budgetType');
-  const modalBudgetType = budgetType === 'rollover' ? 'envelope' : 'tracking';
-
-  const categoryBudgetMenuModal = `${modalBudgetType}-budget-menu`;
-  const categoryNotes = useNotes(category.id);
-
-  const onOpenCategoryBudgetMenu = () => {
-    dispatch(
-      pushModal({
-        modal: {
-          name: categoryBudgetMenuModal,
-          options: {
-            categoryId: category.id,
-            month,
-            onUpdateBudget: amount => {
-              onBudgetAction(month, 'budget-amount', {
-                category: category.id,
-                amount,
-              });
-              showUndoNotification({
-                message: `${category.name} budget has been updated to ${integerToCurrency(amount)}.`,
-              });
-            },
-            onCopyLastMonthAverage: () => {
-              onBudgetAction(month, 'copy-single-last', {
-                category: category.id,
-              });
-              showUndoNotification({
-                message: `${category.name} budget has been set last to month’s budgeted amount.`,
-              });
-            },
-            onSetMonthsAverage: numberOfMonths => {
-              if (
-                numberOfMonths !== 3 &&
-                numberOfMonths !== 6 &&
-                numberOfMonths !== 12
-              ) {
-                return;
-              }
-
-              onBudgetAction(month, `set-single-${numberOfMonths}-avg`, {
-                category: category.id,
-              });
-              showUndoNotification({
-                message: `${category.name} budget has been set to ${numberOfMonths === 12 ? 'yearly' : `${numberOfMonths} month`} average.`,
-              });
-            },
-            onApplyBudgetTemplate: () => {
-              onBudgetAction(month, 'apply-single-category-template', {
-                category: category.id,
-              });
-              showUndoNotification({
-                message: `${category.name} budget templates have been applied.`,
-                pre: categoryNotes,
-              });
-            },
-          },
-        },
-      }),
-    );
-  };
-
-  return (
-    <CellValue
-      binding={binding}
-      type="financial"
-      aria-label={t('Budgeted amount for {{categoryName}} category', {
-        categoryName: category.name,
-      })}
-      {...props}
-    >
-      {({ type, name, value }) =>
-        children?.({
-          type,
-          name,
-          value,
-          onPress: onOpenCategoryBudgetMenu,
-        }) || (
-          <Button
-            variant="bare"
-            style={{
-              ...PILL_STYLE,
-              maxWidth: columnWidth,
-              ...makeAmountGrey(value),
-            }}
-            onPress={onOpenCategoryBudgetMenu}
-            aria-label={t('Open budget menu for {{categoryName}} category', {
-              categoryName: category.name,
-            })}
-          >
-            <View>
-              <PrivacyFilter>
-                <AutoTextSize
-                  key={value}
-                  as={Text}
-                  minFontSizePx={6}
-                  maxFontSizePx={12}
-                  mode="oneline"
-                  style={{
-                    maxWidth: columnWidth,
-                    textAlign: 'right',
-                    fontSize: 12,
-                  }}
-                >
-                  {format(value, type)}
-                </AutoTextSize>
-              </PrivacyFilter>
-            </View>
-          </Button>
-        )
-      }
-    </CellValue>
   );
 }
 
@@ -1184,172 +1051,6 @@ const IncomeGroupHeader = memo(function IncomeGroupHeader({
   );
 });
 
-function IncomeCategoryName({ category, onEdit }) {
-  const sidebarColumnWidth = getColumnWidth({ isSidebar: true, offset: -10 });
-  return (
-    <View
-      style={{
-        flex: 1,
-        justifyContent: 'center',
-        alignItems: 'flex-start',
-        width: sidebarColumnWidth,
-      }}
-    >
-      <Button
-        variant="bare"
-        style={{
-          maxWidth: sidebarColumnWidth,
-        }}
-        onPress={() => onEdit?.(category.id)}
-      >
-        <View
-          style={{
-            flexDirection: 'row',
-            alignItems: 'center',
-            justifyContent: 'flex-start',
-          }}
-        >
-          <Text
-            style={{
-              ...styles.lineClamp(2),
-              width: sidebarColumnWidth,
-              textAlign: 'left',
-              ...styles.smallText,
-            }}
-            data-testid="category-name"
-          >
-            {category.name}
-          </Text>
-          <SvgCheveronRight
-            style={{ flexShrink: 0, color: theme.tableTextSubdued }}
-            width={14}
-            height={14}
-          />
-        </View>
-      </Button>
-    </View>
-  );
-}
-
-function IncomeCategoryCells({ category, month, onBudgetAction }) {
-  const { t } = useTranslation();
-  const format = useFormat();
-  const columnWidth = getColumnWidth();
-  const [budgetType = 'rollover'] = useSyncedPref('budgetType');
-
-  const budgeted =
-    budgetType === 'report' ? trackingBudget.catBudgeted(category.id) : null;
-
-  const balance =
-    budgetType === 'report'
-      ? trackingBudget.catSumAmount(category.id)
-      : envelopeBudget.catSumAmount(category.id);
-
-  return (
-    <View
-      style={{
-        flexDirection: 'row',
-        justifyContent: 'flex-end',
-        alignItems: 'center',
-      }}
-    >
-      {budgeted && (
-        <View
-          style={{
-            width: columnWidth,
-            justifyContent: 'center',
-            alignItems: 'flex-end',
-          }}
-        >
-          <BudgetCell
-            binding={budgeted}
-            type="financial"
-            category={category}
-            month={month}
-            onBudgetAction={onBudgetAction}
-          />
-        </View>
-      )}
-      <CellValue
-        binding={balance}
-        type="financial"
-        aria-label={t('Balance for {{categoryName}} category', {
-          categoryName: category.name,
-        })} // Translated aria-label
-      >
-        {({ type, value }) => (
-          <View>
-            <PrivacyFilter>
-              <AutoTextSize
-                key={value}
-                as={Text}
-                minFontSizePx={6}
-                maxFontSizePx={12}
-                mode="oneline"
-                style={{
-                  width: columnWidth,
-                  justifyContent: 'center',
-                  alignItems: 'flex-end',
-                  textAlign: 'right',
-                  fontSize: 12,
-                  paddingRight: 5,
-                }}
-              >
-                {format(value, type)}
-              </AutoTextSize>
-            </PrivacyFilter>
-          </View>
-        )}
-      </CellValue>
-    </View>
-  );
-}
-
-function IncomeCategoryListItem({
-  month,
-  style,
-  onEdit,
-  onBudgetAction,
-  ...props
-}) {
-  const listItemRef = useRef();
-  const { value: category } = props;
-
-  return (
-    <ListBoxItem
-      textValue={category.name}
-      data-testid="category-row"
-      {...props}
-    >
-      <View
-        style={{
-          height: 50,
-          borderColor: theme.tableBorder,
-          flexDirection: 'row',
-          alignItems: 'center',
-          paddingLeft: 5,
-          paddingRight: 5,
-          zIndex: 1,
-          justifyContent: 'space-between',
-          backgroundColor: 'transparent',
-          borderBottomWidth: 0.5,
-          borderTopWidth: 0.5,
-          opacity: !!category.hidden ? 0.5 : undefined,
-          ...style,
-        }}
-        innerRef={listItemRef}
-      >
-        <IncomeCategoryName category={category} onEdit={onEdit} />
-        <IncomeCategoryCells
-          category={category}
-          month={month}
-          onBudgetAction={onBudgetAction}
-        />
-      </View>
-    </ListBoxItem>
-  );
-}
-
 const ExpenseGroup = memo(function ExpenseGroup({
   type,
   group,
@@ -1565,101 +1266,6 @@ function IncomeGroup({
         />
       </Card>
     </View>
-  );
-}
-
-function IncomeCategoryList({
-  categories,
-  month,
-  onEditCategory,
-  onBudgetAction,
-}) {
-  const { t } = useTranslation();
-  const categoryListData = useListData({
-    initialItems: categories,
-    getKey: category => category.id,
-  });
-  const dispatch = useDispatch();
-
-  const { dragAndDropHooks } = useDragAndDrop({
-    getItems: keys =>
-      [...keys].map(key => ({
-        'text/plain': categoryListData.getItem(key).id,
-      })),
-    renderDropIndicator(target) {
-      return (
-        <DropIndicator
-          target={target}
-          style={{
-            backgroundColor: theme.tableRowBackgroundHighlight,
-            position: 'absolute',
-            left: 2,
-            right: 2,
-            borderRadius: 3,
-            height: 3,
-          }}
-        />
-      );
-    },
-    onReorder(e) {
-      const [key] = e.keys;
-      const categoryIdToMove = key;
-      const categoryGroupId = categoryListData.getItem(key).cat_group;
-      const targetCategoryId = e.target.key;
-
-      if (e.target.dropPosition === 'before') {
-        categoryListData.moveBefore(e.target.key, e.keys);
-
-        dispatch(
-          moveCategory({
-            id: categoryIdToMove,
-            groupId: categoryGroupId,
-            targetId: targetCategoryId,
-          }),
-        );
-      } else if (e.target.dropPosition === 'after') {
-        categoryListData.moveAfter(e.target.key, e.keys);
-
-        const { index: targetIndex } = categoryListData.getItem(e.target.key);
-        const nextToTargetCategory = categoryListData.items[targetIndex + 1];
-
-        dispatch(
-          moveCategory({
-            id: categoryIdToMove,
-            groupId: categoryGroupId,
-            // Due to the way `moveCategory` works, we use the category next to the
-            // actual target category here because `moveCategory` always shoves the
-            // category *before* the target category.
-            // On the other hand, using `null` as `targetId` moves the category
-            // to the end of the list.
-            targetId: nextToTargetCategory?.id || null,
-          }),
-        );
-      }
-    },
-  });
-
-  return (
-    <ListBox
-      aria-label={t('Income categories')}
-      items={categoryListData.items}
-      dragAndDropHooks={dragAndDropHooks}
-    >
-      {category => (
-        <IncomeCategoryListItem
-          key={category.id}
-          value={category}
-          month={month}
-          style={{
-            backgroundColor: monthUtils.isCurrentMonth(month)
-              ? theme.budgetCurrentMonth
-              : theme.budgetOtherMonth,
-          }}
-          onEdit={onEditCategory}
-          onBudgetAction={onBudgetAction}
-        />
-      )}
-    </ListBox>
   );
 }
